refactor(db): 修改部分表的外键为integer类型

+++ backend/base/src/models/hotel_order.rs
@@ -8,13 +8,13 @@ pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
     pub uuid: Uuid,
-    pub hotel_id: i64,
+    pub hotel_id: i32,
     pub begin_date: Date,
     pub end_date: Date,
-    pub hotel_room_type_id: i64,
-    pub person_info_id: i64,
-    pub pay_transaction_id: Option<i64>,
-    pub refund_transaction_id: Option<i64>,
+    pub hotel_room_type_id: i32,
+    pub person_info_id: i32,
+    pub pay_transaction_id: Option<i32>,
+    pub refund_transaction_id: Option<i32>,
     #[sea_orm(column_type = "Decimal(Some((10, 2)))")]
     pub price: Decimal,
     pub create_time: DateTimeWithTimeZone,
